chore: move resize xterm to util
