+++ src/components/v2/appDetails/k8Resource/nodeDetail/NodeDetailTabs/terminal/terminal.utils.tsx
@@ -320,3 +320,15 @@ export default function terminalStripTypeData(elementData) {
             return null
     }
 }
+
+// To fix the scrollbar issue with Xterm in edit mode, we need to restrict the width and height of the xterm-accessibility div as same as xterm-screen div
+// CON: In case of resize, we need to call this function again
+export const restrictXtermAccessibilityWidth = () => {
+    const xtermScreen = document.querySelector('.xterm-screen') as HTMLElement
+    const xtermAccessibility = document.querySelector('.xterm-accessibility') as HTMLElement
+
+    if (xtermScreen && xtermAccessibility) {
+        xtermAccessibility.style.width = xtermScreen.clientWidth + 'px'
+        xtermAccessibility.style.height = xtermScreen.clientHeight + 'px'
+    }
+}
